wip(renderer): rect cells with orphan-anchor body resolution
Rect anchor cells emit their payload bytes as a wide-primary-like cell
spanning the rect's column footprint. Body cells are continuation-like
zero-width placeholders; at emit time they consult the anchor and either
skip (anchor alive) or render as a styled space (anchor orphaned).

+++ uncurses/src/renderer/sync.rs
@@ -63,7 +63,7 @@ impl Renderer {
                 self.back_buf.set_cell(pos, new_cell);
                 // Step over any continuation columns owned by the cell
                 // we just wrote so we don't try to re-copy them.
-                let step = (new_cell.width() as u16).max(1);
+                let step = (new_cell.width()).max(1);
                 x = x.saturating_add(step);
             }
         }
